Handle ICMP identifier as a 32 bit integer
Currently the ICMP identifier in a flow key is a 16-bit integer,
which determines some mismatches when comparing the flow key
value with the conntrack value, and also displays ICMP identifiers
higher than 0x7FFF as negative integers.

With this change the ICMP identifier in the flow key becomes a
32 bit integer, and is converted appropriately when read from or
written to the ICMP packet.

Fixes: MI-910

(cherry picked from commit 15b8051d20929a63787f6b113b8aa1b31d94bfd6)

+++ midonet-util/src/test/java/org/midonet/packets/TestVXLAN.java
@@ -73,7 +73,7 @@ public class TestVXLAN {
         ICMP innerIcmp = (ICMP) innerIp.getPayload();
         assertThat(innerIcmp.getCode(), is(ICMP.CODE_NONE));
         assertThat(innerIcmp.getType(), is(ICMP.TYPE_ECHO_REQUEST));
-        assertThat(innerIcmp.getIdentifier(), is((short)1292));
+        assertThat(innerIcmp.getIdentifier(), is(1292));
 
         assertThat(outerEth.serialize(), is(data));
     }
